requirements.txt
fastapi
uvicorn
numpy
pandas
pillow
scikit-learn
tensorflow
